Move duckdb helper function to duckdb_utils.py
